Don't bail on errors
I really want to catch the error and choose whether to exit based on
whether it was just a failed executable path lookup or something worse,
but I'm still figuring that out. So right now, just assume it's a path
lookup fail and choose not to exit. The user can always type 'exit' or
ctrl+D to get out.