Implement signal handling
The strategy is described in comments but is a little gross. Basically,
there's a global signal handler and it needs to be smart enough to
determine whether the user's intent is to kill a subprocess or the
foreground shell. This is implemented via some shared global vars that
are manipulated via function calls...gross, I know.